# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AquaFlow API",
    "version": "1.0.0",
    "description": "Water Management System API - Comprehensive solution for condominium water consumption tracking, billing, and management.",
    "contact": {
      "name": "AquaFlow Support",
      "email": "support@aquaflow.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.aquaflow.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Use format: Bearer <token>"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "error": {
            "type": "string",
            "description": "Error code or type"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Condominium": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Condominium unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Condominium name"
          },
          "address": {
            "type": "string",
            "description": "Condominium physical address"
          },
          "city": {
            "type": "string",
            "nullable": true,
            "description": "City where the condominium is located"
          },
          "country": {
            "type": "string",
            "nullable": true,
            "description": "Country where the condominium is located"
          },
          "readingDay": {
            "type": "integer",
            "minimum": 1,
            "maximum": 31,
            "nullable": true,
            "description": "Day of the month for meter readings"
          },
          "bankAccount": {
            "type": "string",
            "nullable": true,
            "description": "Bank account for payments"
          },
          "bankAccountHolder": {
            "type": "string",
            "nullable": true,
            "description": "Name of the bank account holder"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the condominium is active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "isActive"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/condominiums/{id}": {
      "get": {
        "summary": "Get condominium details",
        "description": "Retrieve detailed information about a specific condominium",
        "tags": [
          "Condominiums"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Condominium unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Condominium details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Condominium"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - No access to this condominium",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Condominium not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}